+++ python/example_code/neptune/database/neptune_gremlin_query_example.py
@@ -30,7 +30,6 @@ def main():
 
     neptune_client = boto3.client(
         "neptunedata",
-        region_name="us-east-1",
         endpoint_url=NEPTUNE_ENDPOINT,
         config=config
     )
