+++ python/example_code/neptune/database/open_cypher_explain_example.py
@@ -31,7 +31,6 @@ def main():
 
     neptune_client = boto3.client(
         "neptunedata",
-        region_name="us-east-1",
         endpoint_url=NEPTUNE_ENDPOINT,
         config=config
     )
