+++ python/example_code/neptune/hello_neptune.py
@@ -1,0 +1,53 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+# snippet-start:[neptune.python.hello.main]
+import boto3
+from botocore.exceptions import ClientError
+
+
+def describe_db_clusters(neptune_client):
+    """
+    Describes the Amazon Neptune DB clusters using a paginator to handle multiple pages.
+    Raises ClientError with 'ResourceNotFoundException' if no clusters are found.
+    """
+    paginator = neptune_client.get_paginator("describe_db_clusters")
+    clusters_found = False
+
+    for page in paginator.paginate():
+        for cluster in page.get("DBClusters", []):
+            clusters_found = True
+            print(f"Cluster Identifier: {cluster['DBClusterIdentifier']}")
+            print(f"Status: {cluster['Status']}")
+
+    if not clusters_found:
+        raise ClientError(
+            {
+                "Error": {
+                    "Code": "ResourceNotFoundException",
+                    "Message": "No Neptune DB clusters found."
+                }
+            },
+            operation_name="DescribeDBClusters"
+        )
+
+def main():
+    """
+    Main entry point: creates the Neptune client and calls the describe operation.
+    """
+    neptune_client = boto3.client("neptune")
+    try:
+        describe_db_clusters(neptune_client)
+    except ClientError as e:
+        error_code = e.response["Error"]["Code"]
+        if error_code == "ResourceNotFoundException":
+            print(f"Resource not found: {e.response['Error']['Message']}")
+        else:
+            print(f"Unexpected ClientError: {e.response['Error']['Message']}")
+    except Exception as e:
+        print(f"Unexpected error: {str(e)}")
+
+if __name__ == "__main__":
+    main()
+
+# snippet-end:[neptune.python.hello.main]
